misc-coding/ai-research-platform/sub-projects/idea-database/services/content_extractor/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
httpx==0.25.2
structlog==23.2.0
python-multipart==0.0.6
asyncpg==0.29.0
beautifulsoup4==4.12.2
requests==2.31.0
PyPDF2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2
pillow==10.1.0
pytesseract==0.3.10
aiofiles==23.2.1
python-magic==0.4.27
python-dotenv==1.0.0
sqlalchemy==2.0.23
minio==7.2.0
redis==5.0.1
celery==5.3.4
prometheus-client==0.19.0 